- Updated javadoc - Add id for GlobalQuantLayer in Anova
Former-commit-id: 6399980d2b7083ad871e5bc5501d21a4cc2d951e

+++ src/main/java/uk/ac/liv/mzqlib/idmapper/Mzq.java
@@ -1,7 +1,3 @@
-/*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
- */
 
 package uk.ac.liv.mzqlib.idmapper;
 
@@ -95,7 +91,7 @@ public class Mzq {
     }
 
     private double getRetentionWindow(List<Double> massTrace) {
-        return Math.abs(massTrace.get(0).doubleValue() - massTrace.get(2).doubleValue());
+        return Math.abs(massTrace.get(0) - massTrace.get(2));
     }
 
 }
